refactor(cli): move serve command to app/cli/serve.py and collapse main.py
Cuts serve_cmd into app/cli/serve.py and reduces app/main.py to a 9-line
shim that re-exports app from app.cli plus the two render helpers
tests/test_main.py still imports from app.main (_render_task_event,
_truncate_tail). Task 12 will update those tests to canonical imports
and drop the temporary re-export.

After this commit, every command body lives under app/cli/ and the
1108-line monolith is gone.
