+++ app/cli/serve.py
@@ -1,1 +1,101 @@
+"""`atelier serve` command."""
 from __future__ import annotations
+
+import asyncio
+import logging
+from contextlib import asynccontextmanager
+from pathlib import Path
+
+import typer
+
+from app.cli._shared import console
+from app.cli.main import app
+from app.core.atelier import Atelier
+from app.services.api.app import FastApiServer
+from app.services.scheduler import SchedulerDaemon, default_local_zone
+
+
+@app.command(
+    "serve",
+    help="Run the FastAPI HTTP + WebSocket server with the embedded scheduler.",
+)
+def serve_cmd(
+    host: str = typer.Option("127.0.0.1", "--host", help="Bind host."),
+    port: int = typer.Option(
+        8000, "--port", help="Bind port (use 0 for an ephemeral port)."
+    ),
+    reload_interval: float = typer.Option(
+        30.0, "--reload-interval",
+        help="Seconds between schedule store rescans."
+    ),
+    cors_origin: list[str] = typer.Option(
+        [], "--cors-origin",
+        help="Allowed CORS origin (repeatable). Default = '*'."
+    ),
+    log_level: str = typer.Option(
+        "INFO", "--log-level",
+        help="Logging level for the server."
+    ),
+) -> None:
+    import uvicorn
+
+    logging.basicConfig(
+        level=getattr(logging, log_level.upper(), logging.INFO),
+        format="%(asctime)s %(levelname)-7s %(name)s — %(message)s",
+    )
+
+    atelier = Atelier()
+    daemon = SchedulerDaemon(
+        atelier.schedule_store,
+        default_zone=default_local_zone(),
+        default_working_dir=Path.cwd(),
+        reload_interval_seconds=reload_interval,
+    )
+    # The schedules POST/DELETE handlers look here to opportunistically
+    # re-sync the daemon when a schedule is created or removed.
+    atelier.scheduler_daemon = daemon  # type: ignore[attr-defined]
+
+    @asynccontextmanager
+    async def _lifespan(app):
+        await daemon.start()
+        try:
+            yield
+        finally:
+            await daemon.stop()
+
+    cors = list(cors_origin) if cors_origin else None
+    api_app = FastApiServer().create_app(atelier, cors_origins=cors)
+    api_app.router.lifespan_context = _lifespan
+
+    config = uvicorn.Config(
+        api_app,
+        host=host,
+        port=port,
+        log_level=log_level.lower(),
+        lifespan="on",
+    )
+    server = uvicorn.Server(config)
+
+    async def _run() -> None:
+        serve_task = asyncio.create_task(server.serve())
+        # Wait for uvicorn to bind so we can print the actual port.
+        loop = asyncio.get_running_loop()
+        deadline = loop.time() + 10.0
+        while not server.started and loop.time() < deadline:
+            await asyncio.sleep(0.05)
+        actual = port
+        try:
+            if server.servers:
+                actual = server.servers[0].sockets[0].getsockname()[1]
+        except (AttributeError, IndexError, OSError):
+            pass
+        console.print(
+            f"[green]atelier serve[/green] running at "
+            f"[bold]http://{host}:{actual}[/bold]"
+        )
+        await serve_task
+
+    try:
+        asyncio.run(_run())
+    except KeyboardInterrupt:
+        pass
